Changed the place of index_to_slices to more logical place
